requirements.txt
Django==1.9
MySQL-python==1.2.5
oauthlib==0.7.2
PyJWT==1.3.0
python-openid==2.2.5
python-social-auth==0.2.11
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
